Corrections sur le next, mise à jour de l'héritage, tests de contrat cadre, modification de l'implémentation des coûts de production
